pyproject.toml
[project]
name = "adk-mini-cc"
version = "0.1.0"
description = "Modular AI coding agent built on Google ADK"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0",
    "litellm>=1.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "code-index-mcp>=2.13.0",
]

[project.scripts]
adk-mini-cc = "main:main"
